requirements.txt
transformers==4.54.1
torch==2.7.1
gradio==5.9.0
pydantic==2.10.3
langchain==0.3.27
langchain-community==0.3.27
langchain-google-genai==2.1.8
python-dotenv==1.1.1